feat: implement float setFrame support for sub-frame interpolation (Phase 1 - WIP)
Implemented float-based setFrame functionality across all components:

- Property System: Added float setFrame() overload to PropertyBase, Property<T>,
  PropertyGroup, and PropertyArray classes. Updated findKeyframePair() to accept
  float frame parameter with proper sub-frame ratio calculation.

- Core (Composition & Layer): Changed current_frame_ from int to float, added
  float setFrame() overloads with epsilon-based float comparison (FRAME_EPSILON = 0.0001f).
  Updated getCurrentFrame() to return float, added getCurrentFrameInt() for compatibility.

- Source System: Added float setFrame() overloads to all source classes
  (LayerSource, CompositionSource, VideoSource, SequenceSource, ShapeSource).
  Sources that require integer frames (Video, Sequence) use static_cast<int>(frame).

- Player: Changed target_frame_ from int to float, added float setFrame() and
  constrainFrame() overloads. Updated playback logic to use float frame arithmetic.

Backward Compatibility: All integer setFrame(int) methods now call the float
version via setFrame(static_cast<float>(frame)), ensuring existing code continues
to work.

This enables sub-frame interpolation (e.g., frame 10.5 for 15fps content at 30fps
app framerate) while maintaining full backward compatibility.

Related: FRAME_INTERPOLATION_ANALYSIS.md, TIME_MIGRATION_ROADMAP.md Phase 1

+++ src/prop/ofxAEKeyframe.h
@@ -173,7 +173,7 @@ struct KeyframePair {
 };
 
 template<typename T>
-KeyframePair<T> findKeyframePair(const std::map<int, Keyframe::Data<T>>& keyframes, int frame) {
+KeyframePair<T> findKeyframePair(const std::map<int, Keyframe::Data<T>>& keyframes, float frame) {
 	KeyframePair<T> result;
 	
 	if (keyframes.empty()) {
@@ -190,7 +190,7 @@ KeyframePair<T> findKeyframePair(const std::map<int, Keyframe::Data<T>>& keyfram
 		return result;
 	}
 	
-	auto upper = keyframes.upper_bound(frame);
+	auto upper = keyframes.upper_bound(static_cast<int>(frame));
 	
 	if (upper == keyframes.begin()) {
 		auto first = keyframes.begin();
@@ -221,7 +221,7 @@ KeyframePair<T> findKeyframePair(const std::map<int, Keyframe::Data<T>>& keyfram
 	result.frame_b = keyframe_b_it->first;
 	
 	if (result.frame_b != result.frame_a) {
-		result.ratio = static_cast<float>(frame - result.frame_a) / static_cast<float>(result.frame_b - result.frame_a);
+		result.ratio = (frame - static_cast<float>(result.frame_a)) / static_cast<float>(result.frame_b - result.frame_a);
 	}
 	else {
 		result.ratio = 0.0f;
