+++ src/source/ofxAESequenceSource.h
@@ -11,7 +11,7 @@ class SequenceSource : public LayerSource
 public:
 	void accept(Visitor &visitor) override;
 	bool load(const std::filesystem::path &filepath) override;
-	bool setFrame(int frame) override;
+	bool setFrame(float frame) override;
 	void draw(float x, float y, float w, float h) const override { if(texture_) (*texture_)->draw(x,y,w,h); }
 	float getWidth() const override { return pool_.empty() ? 0.f : pool_[0].getWidth(); }
 	float getHeight() const override { return pool_.empty() ? 0.f : pool_[0].getHeight(); }
